# doc/swagger.yml
openapi: "3.0.1"
info: 
  title: "API MOSAIQUAGE"
  description: "Documentation de l'API pour le controle et les retouches de mosaiquage."
servers: 
  - url: "http://localhost:8081"
    description: "Serveur de dev"
...
paths:
...
  '/{idBranch}/vectors':
    get:
      tags:
        - vector
      summary: "Récupération des métadonnées de tous les vecteurs annexes d'une branche"
      description: ""
      parameters:
        - in: path
          name: idBranch
          description: l'identifiant de la branche
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
  '/{idBranch}/vector':
    get:
      tags:
        - vector
      summary: "Récuperation d'un vecteur annexe à partir de son nom"
      description: ""
      parameters:
        - in: path
          name: idBranch
          description: l'identifiant de la branche
          required: true
          schema:
            type: integer
        - in: query
          name: name
          description: le nom de la couche vecteur
          required: true
          schema:
            type: string
          example: Remarques
      responses:
        '200':
          description: OK
  
    post:
      tags:
        - vector
      summary: "Ajout d'un vecteur annexe en base"
      description: "Ajout d'un geojson en base. Géometrie accepté: Point / MultiPoint / LineString / MultiLineString / Polygon / MultiPolygon"
      parameters:
        - in: path
          name: idBranch
          description: l'identifiant de la branche
          required: true
          schema:
            type: integer
        - in: query
          name: name
          description: le nom de la couche vecteur
          required: true
          schema:
            type: string
      requestBody:
        description: couche vecteur annexe à ajouter (GeoJSON).
        required: true
        content:
          application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                    value: FeatureCollection
                  features:
                    type: array
                    items:
                      type: object
                      properties:
                        type:
                          type: string
                          example: Feature
                        properties:
                          type: object
                          properties:
                            prop_as_you_want:
                              type: sting
                              example: as_you_want
                        geometry:
                          type: object
                          properties:
                            type:
                              type: string
                              enum:
                                - Point
                                - MultiPoint
                                - LineString
                                - MultiLineString
                                - Polygon
                                - MultiPolygon
                            coordinates:
                              type: array
                              example: [230747.32, 6759652.52]

      responses:
        '200':
          description: OK
  '/vector':
    get:
      tags:
        - vector
      summary: "Récupération d'un vecteur annexe donné à partir de son identifiant"
      description: ""
      parameters:
        - in: query
          name: idVector
          description: l'identifiant du vecteur
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK

    delete:
      tags:
        - vector
      summary: "Destruction d'un vecteur donné"
      description: ""
      parameters:
        - in: query
          name: idVector
          description: l'identifiant du vecteur annexe à detruire
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
  '/{idRemarksVector}/feature':
    put:
      tags:
        - vector
      summary: "Ajout d'un ponctuel sur une couche 'Remarques'"
      description: "Ne permet d'ajouter un ponctuel que sur une couche 'Remarques'."
      parameters:
        - in: path
          name: idRemarksVector
          description: l'identifiant de la couche 'Remarques' à modifier
          required: true
          schema:
            type: integer
        - in: query
          name: x
          description: coordonnées x
          required: true
          schema:
            type: number
        - in: query
          name: y
          description: coordonnées y
          required: true
          schema:
            type: number
        - in: query
          name: comment
          description: commentaire
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK

    delete:
      tags:
        - vector
      summary: "Suppression d'un ponctuel sur une couche 'Remarques'"
      description: "Ne permet de supprimer un ponctuel que sur une couche 'Remarques'."
      parameters:
        - in: path
          name: idRemarksVector
          description: l'identifiant de la couche 'Remarques' à modifier
          required: true
          schema:
            type: integer
        - in: query
          name: id
          description: id de la feature à supprimer
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK

# Alerts
  '/vector/{idFeature}?status={status}':
    put:
      tags:
        - vector
      summary: "Modification de l'attribut 'status' d'une entité donnée"
      description: ""
      parameters:
        - in: path
          name: idFeature
          description: l'identifiant de la feature à modifier
          required: true
          schema:
            type: integer
        - in: path
          name: status
          description: nouvelle valeur
          required: true
          schema:
            type: boolean
      responses:
        '200':
          description: OK
  '/vector/{idFeature}?comment={comment}':
    put:
      tags:
        - vector
      summary: "Modification de l'attribut 'comment' d'une entité donnée"
      description: ""
      parameters:
        - in: path
          name: idFeature
          description: l'identifiant de la feature à modifier
          required: true
          schema:
            type: integer
        - in: path
          name: comment
          description: nouvelle valeur
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK

# WMTS
...