# routes/vector.js
const router = require('express').Router();
const {
  param, query, body, oneOf,
} = require('express-validator');
const GJV = require('geojson-validation');
const branch = require('../middlewares/branch');
const vector = require('../middlewares/vector');
const validateParams = require('../paramValidation/validateParams');
const createErrMsg = require('../paramValidation/createErrMsg');
const pgClient = require('../middlewares/pgClient');
const returnMsg = require('../middlewares/returnMsg');

// Encapsulation des informations du requestBody dans une nouvelle clé 'keyName' ("body" par defaut)
function encapBody(req, res, next) {
  let keyName = 'body';
  if (this.keyName) { keyName = this.keyName; }
  if (JSON.stringify(req.body) !== '{}') {
    const requestBodyKeys = Object.keys(req.body);
    req.body[keyName] = JSON.parse(JSON.stringify(req.body));
    for (let i = 0; i < requestBodyKeys.length; i += 1) {
      delete req.body[requestBodyKeys[i]];
    }
  }
  next();
}

const vectorToSave = [
  body('json')
    .if(body('json').exists())
    .custom(GJV.isGeoJSONObject)
    .withMessage(createErrMsg.invalidBody('objet GeoJSON'))
    .if(body('json').exists())
    .custom(GJV.isFeatureCollection)
    .withMessage(createErrMsg.invalidBody('featureCollection')),
  body('json.type')
    .if(body('json').exists())
    .exists().withMessage(createErrMsg.missingParameter('type'))
    .isIn(['FeatureCollection'])
    .withMessage(createErrMsg.invalidParameter('type')),
  body('json.features.*.geometry')
    .custom((value) => (
      GJV.isPoint(value)
      || GJV.isMultiPoint(value)
      || GJV.isLineString(value)
      || GJV.isMultiLineString(value)
      || GJV.isPolygon(value)
      || GJV.isMultiPolygon(value)
    ))
    .withMessage(createErrMsg.InvalidEntity('data.features.*.geometry',
      '(Point, MultiPoint, LineString, MultiLineString, Polygon ou MultiPolygone)')),
];

router.get('/:idBranch/vectors',
  pgClient.open,
  branch.getBranches.bind({ column: 'id' }),
  [
    param('idBranch')
      .exists().withMessage(createErrMsg.missingParameter('idBranch'))
      .custom((value, { req }) => req.result.getBranches.includes(Number(value)))
      .withMessage(createErrMsg.invalidParameter('idBranch')),
  ],
  validateParams,
  vector.getVectors,
  pgClient.close,
  returnMsg);

router.get('/:idBranch/vector',
  pgClient.open,
  branch.getBranches.bind({ column: 'id' }),
  param('idBranch')
    .exists().withMessage(createErrMsg.missingParameter('idBranch'))
    .custom((value, { req }) => req.result.getBranches.includes(Number(value)))
    .withMessage(createErrMsg.invalidParameter('idBranch')),
  validateParams,
  vector.getVectors.bind({ column: 'name' }),
  query('name')
    .exists().withMessage(createErrMsg.missingParameter('name'))
    .custom((value, { req }) => req.result.getVectors.includes(value))
    .withMessage(createErrMsg.invalidParameter('name')),
  validateParams,
  vector.getVector,
  pgClient.close,
  returnMsg);

router.post('/:idBranch/vector', encapBody.bind({ keyName: 'json' }),
  pgClient.open,
  branch.getBranches.bind({ column: 'id' }),
  vector.getVectors.bind({ column: 'name' }),
  [
    param('idBranch')
      .exists().withMessage(createErrMsg.missingParameter('idBranch'))
      .custom((value, { req }) => req.result.getBranches.includes(Number(value)))
      .withMessage(createErrMsg.invalidParameter('idBranch')),
    query('name')
      .exists().withMessage(createErrMsg.missingParameter('name'))
      .custom((value, { req }) => !req.result.getVectors.includes(value))
      .withMessage(createErrMsg.invalidParameter('name')),
    body('json')
      .exists().withMessage(createErrMsg.missingBody),
    ...vectorToSave,
  ],
  validateParams,
  vector.postVector,
  pgClient.close,
  returnMsg);

router.get('/vector',
  pgClient.open,
  vector.getVectors.bind({ column: 'id' }),
  query('idVector')
    .exists().withMessage(createErrMsg.missingParameter('idVector'))
    .custom((value, { req }) => req.result.getVectors.includes(Number(value)))
    .withMessage(createErrMsg.invalidParameter('idVector')),
  validateParams,
  vector.getVector,
  pgClient.close,
  returnMsg);

router.delete('/vector',
  pgClient.open,
  vector.getVectors.bind({ column: 'id' }),
  query('idVector')
    .exists().withMessage(createErrMsg.missingParameter('idVector'))
    .custom((value, { req }) => req.result.getVectors.includes(Number(value)))
    .withMessage(createErrMsg.invalidParameter('idVector')),
  validateParams,
  vector.deleteVector,
  pgClient.close,
  returnMsg);

router.put('/vector/:idFeature',
  pgClient.open,
  vector.getFeatures.bind({ column: 'id' }),
  param('idFeature')
    .exists().withMessage(createErrMsg.missingParameter('idFeature'))
    .custom((value, { req }) => req.result.getFeatures.includes(Number(value)))
    .withMessage(createErrMsg.invalidParameter('idFeature')),
  oneOf([
    query('status')
      .exists().withMessage(createErrMsg.missingParameter('status')),
    query('comment')
      .exists().withMessage(createErrMsg.missingParameter('comment')),
  ]),
  validateParams,
  vector.updateAlert,
  pgClient.close,
  returnMsg);

router.put('/:idRemarksVector/feature',
  pgClient.open,
  vector.getVectors.bind({ column: 'id', selection: { column: 'name', value: 'Remarques' } }),
  param('idRemarksVector')
    .exists().withMessage(createErrMsg.missingParameter('idRemarksVector'))
    .custom((value, { req }) => req.result.getVectors.includes(Number(value)))
    .withMessage(createErrMsg.invalidParameter('idRemarksVector')),
  query('x')
    .exists().withMessage(createErrMsg.missingParameter('x')),
  query('y')
    .exists().withMessage(createErrMsg.missingParameter('y')),
  query('comment')
    .exists().withMessage(createErrMsg.missingParameter('comment')),
  validateParams,
  vector.addRemark,
  pgClient.close,
  returnMsg);

router.delete('/:idRemarksVector/feature',
  pgClient.open,
  vector.getVectors.bind({ column: 'id', selection: { column: 'name', value: 'Remarques' } }),
  param('idRemarksVector')
    .exists().withMessage(createErrMsg.missingParameter('idRemarksVector'))
    .custom((value, { req }) => req.result.getVectors.includes(Number(value)))
    .withMessage(createErrMsg.invalidParameter('idRemarksVector')),
  query('id')
    .exists().withMessage(createErrMsg.missingParameter('id')),
  validateParams,
  vector.delRemark,
  pgClient.close,
  returnMsg);

module.exports = router;